async: remove node from db only when it isn't available in the API
This fixes a longstanding issue with OVH where mita requests a node to
be destroyed, OVH ACKs, but since the removal happens asynchronously,
there are times where the nodes sticks around. With this change we are
only removing it from the database when the API reports it is no longer
there.

Signed-off-by: Alfredo Deza <adeza@redhat.com>