controllers: verify node is idle one last time
Talking to Jenkins and removing a node is a lot faster than talking to
the Cloud provider, waiting for a response and then trying to remove the
node from Jenkins.

This makes it so that Jobs aren't picking up a node in between those
actions and then getting the node destroyed. By talking to Jenkins
first, and asking if the node is really idle, we reduce the window of
opportunity for this to happen

Signed-off-by: Alfredo Deza <adeza@redhat.com>