mita/controllers/nodes: Increasing timeout to 40mn
A recent study on the CI shows that on 300 jobs:
- the average time between two jobs submisson is 25mn
- the standard deviation between two jobs submission is 50mn.
- 187 build hosts were used
- the average build takes 43mn

Since the CI is now properly using the ccache to speed the build, a
build can last 14mn when the local cache is hit.

With a timeout of 20mn, the build nodes dies before a new PR is incoming
loosing the built cache that could have been reused in we had waited a
little bit longer before destroying it.

The goal of this commit is to increase the chance that two recently
submitted PRs are sharing the same cache running on the node. As a
result, the job duration will tend to 14mn.

Signed-off-by: Erwan Velu <erwan@redhat.com>