prod_nodes: Create centos7_gigantic node
I ran some stats gathering on shaman and discovered the CentOS7 braggi builders were extremely underused.  So I reimaged 3 of them to CentOS8.  Then, yesterday, Yuri had a bunch of PR testing to do for the next Nautilus release which involved a lot of CentOS7 builds needing to be created.  I know we are trying to lessen our OVH usage but I think CentOS7 is the best distro for us to dynamically spin up nodes for since it will be the least used of them (Ubuntu/CentOS8) all.

Signed-off-by: David Galloway <dgallowa@redhat.com>